// ra150396/frontend/src/app/models/error-messages.model.ts
export class ErrorMessages {
    name: string = '';
    lastName: string = '';
    email: string = '';
    profession: string = '';
    username: string = '';
    password: string = '';
    passwordConfirmation: string = '';
    gender: string = '';
    jmbg: string = '';
    imgPath: string = '';
    securityQuestion: string = '';
    securityAnswer: string = '';
    hasErrors: boolean = false;
};